gimbal: fix input selection
Once we have valid input we should use that and not overwrite our
update_result with the result from other inputs, otherwise the
automatic selection does not actually work.

This behavior means that only the first update will be used if there are
several sources updating in the same cycle.


Former-commit-id: 9b35b680f64be1dd812f19bcc0b9d911e35b0594