Commander: remove `COM_POS_FS_DELAY`
A user configurable delay for the internal `vehicle_local_position` seems confusing in my eyes. It's a different timeout for fixed-wing and multirotor which might have made sense earlier but not really anymore since the topic is constantly published by the estimator and not expected to time out on either vehicle type and the parameter description is also misleading because it's outdated.


Former-commit-id: d416cd2a6cfc34a59cf8f85c25e7c90286049ba7